moved iterative props to GenericAlg, exposed iterative_props
